openai-agents-sdk/thebostonwrongs/requirements.txt
openai-agents
openai-agents[litellm]
python-dotenv
aiohttp
